track any stderr output from spawned child processes
added a new method to base Job class @spawn.  this method will track any stderr
output generated by the child process.  if the process returns an error code, it
will output the stderr output to the mimelog.  this makes error conditions a
little easier to resolve but looking at the logs.