layer searchable text behind original pdf
i screwed up an earlier git commit so this commit represents two large chunks
of work:
- Update PageGenerator job.  now creates a visually blank searchable
pdf
- Add PdfLayer job. layers blank searchable pdf under the original pdf
page.  this results in a searchable pdf page that is visually identical
to the original pdf

These changes mean that there are a number of "versions" of pdfs for
each page stored in redis.  Some of these will exist is redis simultaneously.
In order to ensure that there wasn't any accidental overwriting of these
files due to key collision I am using distinct file extensions:
- *.pdf: the original pdf, the original pdf pages or the final pdf
- *.spdf: visually blank searchable page
- *.lpdf: layered pdf page.  original pdf over blank searchable page