Support creating searchable pdfs of arbitrary sizes
The previous version was stuck on making searching PDFs that
were letter sized because it used a stock blank.jpg with letter
sized size/density.  Mimeograph now determines the size the source
page by parsing the hOCR page and then generates a blank image
to fit it.

- utils.coffee: log.debug now using stringify.  this is helpful for
debugging
- mimeograph.coffee: PageGenerator is a pretty big beast now.  it now
handles a multistep process to create the blank image used to create
the searchable blank pdf.  i decided against breaking this out into
separate jobs because the workflow because i don't think it was waranted:
this mini-flow is pretty isolated and broken out into a series of
callbacks.
- package.json: added a new dependency to help parse the hOCR xml.
converting the xml to json was the best of all of the nasty options.
although i would prefer a straight forward xpath expression, the depdency
we are using "xml2js" is pure JS.