Make jobs a little more cautious about potential error conditions
I have found a few "orphaned" jobs in production.  These appear
to be due to jobs that complete normally but in actuality
do have not executed responsibilities as expected.

I think this is largely because of the following line in Job.spawn:

    if code

Nodejs will return a null code if the child process does not exit
normally (e.g. tool has a seg fault).  If this is the case the job
will think its process has exited normaly and continue on its mary
way.  Unfortunatly the code in Mimeograph that sits "between" jobs
such as @pdf, @split etc ... will most likely error out.

I solved this by changing the line above to:

unless code == 0

I also added a number of path.exists checks in the various jobs.
This logic needs to be broken out into a separate method for Job.